requirements.txt
asgiref==3.7.2
asttokens==2.2.1
Babel==2.12.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.2
django-autoslug==1.9.9
django-ckeditor==6.6.1
django-extensions==3.2.3
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
django-smart-selects==1.6.0
executing==1.2.0
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
phonenumbers==8.13.15
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
wcwidth==0.2.6
